dcp_requirement.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.3.0
djangorestframework==3.11.0
Faker==4.0.3
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
